+++ src/Statement/StatementList.php
@@ -4,6 +4,7 @@ namespace Wikibase\DataModel\Statement;
 
 use InvalidArgumentException;
 use Traversable;
+use Wikibase\DataModel\Claim\Claim;
 use Wikibase\DataModel\Claim\Claims;
 use Wikibase\DataModel\Entity\PropertyId;
 use Wikibase\DataModel\Reference;
@@ -106,7 +107,7 @@ class StatementList implements \IteratorAggregate, \Comparable, \Countable {
 		$qualifiers = is_array( $qualifiers ) ? new SnakList( $qualifiers ) : $qualifiers;
 		$references = is_array( $references ) ? new ReferenceList( $references ) : $references;
 
-		$statement = new Statement( $mainSnak, $qualifiers, $references );
+		$statement = new Statement( new Claim( $mainSnak, $qualifiers ), $references );
 		$statement->setGuid( $guid );
 
 		$this->addStatement( $statement );